edit permissions

+++ src/base/views.py
@@ -62,9 +62,11 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
-@login_required(login_url='/login')
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
+    if request.user != room.host:
+        return HttpResponse("You don't have permissions to complete your action")
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -74,12 +76,12 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-@login_required(login_url='/login')
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
-    if request.user != room.user:
+    if request.user != room.host:
         return HttpResponse("You don't have permissions to complete your action")
 
     if request.method == 'POST':
@@ -91,9 +93,11 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-@login_required(login_url='/login')
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("You don't have permissions to complete your action")
 
     if request.method == 'POST':
         room.delete()
